refactor: draw activity test rows from a mock-row factory
create-activity-row.ts baked its faker defaults inline instead of
drawing them from a create-mock-* factory, the only entity in the
package to skip that layer. Add
test-utils/factories/create-mock-activity-row.ts and have the
composite build its insert values from it, matching
create-avatar-row.ts / create-mock-avatar-row.ts.
