+++ libs/data/active-avatar/src/test-utils/create-activity-row.ts
@@ -1,7 +1,6 @@
-import { faker } from '@faker-js/faker';
-import { createId } from '@paralleldrive/cuid2';
 import type { Activities, DB } from '@vers/db';
 import type { Insertable, Kysely, Selectable } from 'kysely';
+import { createMockActivityRow } from './factories/create-mock-activity-row';
 
 interface CreateActivityRowData extends Partial<Insertable<Activities>> {
   readonly avatarId: string;
@@ -9,30 +8,13 @@ interface CreateActivityRowData extends Partial<Insertable<Activities>> {
 
 /**
  * Seeds an activity head row for an existing avatar — the live-run query tests need a real
- * `status = 'active'` row. Hash and content fields default to opaque strings: the query under
- * test reads status and ownership, never the chain.
+ * `status = 'active'` row.
  */
 export function createActivityRow(
   db: Kysely<DB>,
   data: Readonly<CreateActivityRowData>,
 ): Promise<Selectable<Activities>> {
-  const startHash = faker.string.hexadecimal({ casing: 'lower', length: 64, prefix: '' });
+  const row = createMockActivityRow(data);
 
-  return db
-    .insertInto('activities')
-    .values({
-      buildSnapshot: { level: 1, xp: 0 },
-      contentVersion: '0.0.0-dev',
-      encounterNode: { difficulty: 1 },
-      id: `act_${createId()}`,
-      lastHash: startHash,
-      scopeId: faker.string.alphanumeric({ casing: 'lower', length: 8 }),
-      scopeType: 'world_map_node',
-      seed: faker.string.hexadecimal({ casing: 'lower', length: 32, prefix: '' }),
-      simVersion: '0.0.0-dev',
-      startHash,
-      ...data,
-    })
-    .returningAll()
-    .executeTakeFirstOrThrow();
+  return db.insertInto('activities').values(row).returningAll().executeTakeFirstOrThrow();
 }
